marvellore-ci/requirements.txt
flask==3.0.3
pdfplumber==0.11.4
requests==2.32.3
PyGithub==2.3.0
python-dotenv==1.0.1
pyngrok==7.2.0
APScheduler==3.10.4
openai==1.35.0
